Fix infinite reload loop
Rebuilding the menu by executing `xcrun simctl list -j devices` triggers the directory watcher to `Developer/CoreSimulator/Devices`, the completion of which rebuilds the menu...

It's a simple workaroud to stop the watcher and restart it after the command finishes executing. It might be more appropriate to stop relying on `xcrun simctl` though.

This should fix #16 and #19.