Fix device parser
On Xcode 10.1 there’s a new property on the device json which is not a String:
 "isAvailable" : true,

This was causing the parser to fail due the [[String:String]] casting.

I changed it to [[String:AnyObject]] and added String casts on the if let.

Ideally this would be solved by codable, but it would require a bigger refactor. I might give it a shot later but I wanted a quick fix to keep using this ASAP :)