Increase automatic reload time delay
Since when running or starting a simulator multiple files are changed on the disk, the file notifier was sending multiple reload requests that were forcing the CPU to do unecessary refreshes. This workaround increases the delay time to 5 seconds for file changes while keeping the standard (1 second) for user refresh.

Ideally this would be solved by lazy loading some information only when the user opens the menu.